# lib/api-zod/src/generated/types/createExpenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseInput {
  category: string;
  amount: number;
  description?: string | null;
}