# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  sku?: string | null;
  category?: string | null;
  quantity: number;
  cost_price: number;
  sale_price: number;
  low_stock_threshold?: number | null;
  created_at: string;
}