# lib/api-zod/src/generated/types/purchaseWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseItem } from "./purchaseItem";
import type { PurchaseWithItemsPaymentType } from "./purchaseWithItemsPaymentType";
import type { PurchaseWithItemsStatus } from "./purchaseWithItemsStatus";

export interface PurchaseWithItems {
  id: number;
  invoice_no: string;
  supplier_name?: string | null;
  supplier_id?: number | null;
  payment_type: PurchaseWithItemsPaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: PurchaseWithItemsStatus;
  notes?: string | null;
  created_at: string;
  items: PurchaseItem[];
}